+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.2.0
 
  - `resizeMode` prop added, now defaults to `contain`
+ - `customStyles` prop added, allows for customization of the player.